first shot at eval graph

+++ src/de/jose/pgn/MoveNode.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of the Jose Project
  * see http://jose-chess.sourceforge.net/
- * (c) 2002-2006 Peter Schäfer
+ * (c) 2002-2006 Peter Schï¿½fer
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -42,7 +42,7 @@ public class MoveNode
 	protected Move move;
     protected long hashKey;
 	protected int moveCountLen;
-	public double[] engineValue;
+	public float[] engineValue;
 
     protected static MoveNodeFormatter formatter = new MoveNodeFormatter();
 	protected static SAXMoveFormatter saxFormatter = new SAXMoveFormatter();
