+++ src/de/jose/view/BoardView.java
@@ -1,7 +1,7 @@
 /*
  * This file is part of the Jose Project
  * see http://jose-chess.sourceforge.net/
- * (c) 2002-2006 Peter Schäfer
+ * (c) 2002-2006 Peter Schï¿½fer
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -80,7 +80,7 @@ abstract public class BoardView
 	protected boolean showAnimationHints;
 
 	//	score for eval bar
-	protected double[] eval = null;
+	protected float[] eval = null;
 
 	public BoardView(IBoardAdapter theBoard)
 	{
